remove(log): ControlBar 컴포넌트 제거 (#108)
