+++ app/(main)/log/page.tsx
@@ -1,12 +1,19 @@
 'use client'
 import { Group, Pagination } from '@mantine/core'
+import Image from 'next/image'
 import { useRouter } from 'next/navigation'
 import { useState } from 'react'
 
 import Breadcrumb from '@/components/common/Breadcrumb'
-import ControlBar from '@/components/domain/vehicle/ControlBar'
+import ExcelButton from '@/components/common/Button/ExcelButton'
+import LinkButton from '@/components/common/Button/LinkButton'
+import ControlLayout from '@/components/common/ControlLayout'
+import SearchInput from '@/components/common/Input/SearchInput'
+import Modal from '@/components/common/Modal'
+import { ModalMessageType } from '@/components/common/Modal/types'
 import ListHeader from '@/components/domain/vehicle/ListHeader'
 import { LOG_TITLES } from '@/constants/listHeader'
+import { useSearchSingleVehicle } from '@/hooks/useSearchSingleVehicle'
 
 import ListItem from './components/ListItem/index'
 import { useLogData } from './hooks/useLogData'
@@ -17,6 +24,8 @@ const LogPage = () => {
     const router = useRouter()
     const [activePage, setActivePage] = useState(1)
     const { logData, isLoading, error } = useLogData(activePage - 1)
+    const { modalMessage, isOpen, searchTerm, handleVehicleSearch, handleSearchChange, closeModal } =
+        useSearchSingleVehicle()
 
     const handleExcelButtonClick = () => {
         if (!logData?.content) return
@@ -37,13 +46,40 @@ const LogPage = () => {
     return (
         <div className={styles.container}>
             <Breadcrumb type={'운행기록'} />
+
             <div className={styles.contents}>
-                <ControlBar onExcelButtonClick={handleExcelButtonClick} />
+                <ControlLayout
+                    control={
+                        <SearchInput
+                            icon='/icons/search-icon.svg'
+                            value={searchTerm}
+                            onChange={handleSearchChange}
+                            onSubmit={handleVehicleSearch}
+                        />
+                    }
+                    primaryButton={<ExcelButton onClick={handleExcelButtonClick} />}
+                    secondaryButton={
+                        <LinkButton href={'/log/register'}>
+                            <div className={styles.linkButton}>
+                                <Image src='/icons/white-add-icon.svg' alt='add' width={18} height={18} />
+                                등록
+                            </div>
+                        </LinkButton>
+                    }
+                />
+                <Modal
+                    isOpen={isOpen}
+                    message={modalMessage as ModalMessageType}
+                    variant={{ variant: 'alert', confirmButton: '확인' }}
+                    onClose={closeModal}
+                />
+
                 <ListHeader headerTitles={LOG_TITLES} />
                 {logData?.content.map((log) => (
                     <ListItem key={log.id} data={log} onClick={() => handleItemClick(log.id)} />
                 ))}
             </div>
+
             <div className={styles.pagination}>
                 <Pagination.Root total={logData?.totalPages || 1} value={activePage} onChange={setActivePage}>
                     <Group gap={5} justify='center'>
